+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Coming npm version
 
+## npm v0.2.10
+Add project information section in header.
+Aesthetic changes in gantt table.
 ## npm v0.2.9
 Changes style, Adds assignee on resource view